# openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Uganda-first professional + dating network. Envelope: `{code:1|0, message, data}`. Auth: Bearer JWT.",
    "title": "LinkUp API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/wallet": {
      "get": {
        "operationId": "get_wallet.get_wallet",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/wallet",
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/wallet/transactions": {
      "get": {
        "operationId": "get_wallet.transactions",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /api/wallet/transactions",
        "tags": [
          "api"
        ]
      }
    },
...
    "/v1/wallet/balance": {
      "get": {
        "operationId": "get_v1_wallet.balance",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/wallet/balance",
        "tags": [
          "wallet"
        ]
      }
    },
    "/v1/wallet/topup": {
      "post": {
        "operationId": "post_v1_wallet.topup",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /v1/wallet/topup",
        "tags": [
          "wallet"
        ]
      }
    },
    "/v1/wallet/topup/{tx_ref}/verify": {
      "get": {
        "operationId": "get_v1_wallet.verify_topup",
        "parameters": [
          {
            "in": "path",
            "name": "tx_ref",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/wallet/topup/{tx_ref}/verify",
        "tags": [
          "wallet"
        ]
      },
      "post": {
        "operationId": "post_v1_wallet.verify_topup",
        "parameters": [
          {
            "in": "path",
            "name": "tx_ref",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "POST /v1/wallet/topup/{tx_ref}/verify",
        "tags": [
          "wallet"
        ]
      }
    },
    "/v1/wallet/transactions": {
      "get": {
        "operationId": "get_v1_wallet.transactions",
        "responses": {
          "200": {
            "description": "Success \u2014 `{code:1, message, data}`"
          },
          "400": {
            "description": "Error \u2014 `{code:0, message}`"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "GET /v1/wallet/transactions",
        "tags": [
          "wallet"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
...
}